+++ bin/irregular/rebuild_mongo.dart
@@ -559,13 +559,13 @@ Future<void> main() async {
 
   /// ISONE
   // await recreateCompetitiveOffersIsone();
-  // await recreateDaBindingConstraintsIsone();
+  await recreateDaBindingConstraintsIsone();
   // await recreateDaLmpHourlyIsone();
   // await recreateDaCongestionCompactIsone();
   // await recreateDaDemandBid();
   // await recreateDaEnergyOffersIsone();
   // await recreateFwdResAuctionResults();
-  await recreateIsoneHistoricalBtmSolar();
+  // await recreateIsoneHistoricalBtmSolar();
   // await insertMaskedAssetIdsIsone();
   // await recreatePtidTableIsone();
   // await recreateRegulationRequirementIsone();